+++ actionview/test/template/test_case_test.rb
@@ -369,7 +369,7 @@ module ActionView
     include ::Capybara::Minitest::Assertions
 
     def page
-      Capybara.string(document_root_element)
+      Capybara.string(rendered)
     end
 
     test "document_root_element can be configured to utilize Capybara" do
@@ -382,15 +382,16 @@ module ActionView
     end
   end
 
-  class RenderedMethodMissingTest < ActionView::TestCase
-    test "rendered delegates methods to the String" do
+  class RenderedViewContentTest < ActionView::TestCase
+    test "#rendered inherits from String" do
       developer = DeveloperStruct.new("Eloy")
 
       render "developers/developer", developer: developer
 
+      assert_kind_of String, rendered
       assert_kind_of String, rendered.to_s
       assert_equal developer.name, rendered
-      assert_match rendered, /#{developer.name}/
+      assert_match(/#{developer.name}/, rendered)
       assert_includes rendered, developer.name
     end
   end