+++ actionview/CHANGELOG.md
@@ -1,8 +1,13 @@
+*   Rename `ActionView::TestCase::Behavior::{Content,RenderedViewContent}`
+
+    *Sean Doyle*
+
 *   Fix detection of required strict locals
 
     Further fix `render @collection` compatibility with strict locals
 
     *Jean Boussier*
+
 
 ## Rails 7.1.2 (November 10, 2023) ##